# lib/api-zod/src/generated/types/job.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Job {
  id: number;
  title: string;
  company: string;
  location: string;
  /** e.g. Full-time, Part-time, Contract, Remote */
  type: string;
  /** e.g. Engineering, Design, Marketing */
  category: string;
  /** Salary range e.g. $80k - $120k */
  salary?: string;
  description: string;
  /** Company logo URL or initials */
  logo?: string;
  postedAt: Date;
  featured: boolean;
}