# lib/db/src/schema/jobs.ts
import { pgTable, text, serial, boolean, timestamp } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const jobsTable = pgTable("jobs", {
  id: serial("id").primaryKey(),
  title: text("title").notNull(),
  company: text("company").notNull(),
  location: text("location").notNull(),
  type: text("type").notNull(),
  category: text("category").notNull(),
  salary: text("salary"),
  description: text("description").notNull(),
  logo: text("logo"),
  postedAt: timestamp("posted_at").notNull().defaultNow(),
  featured: boolean("featured").notNull().default(false),
});

export const insertJobSchema = createInsertSchema(jobsTable).omit({ id: true });
export type InsertJob = z.infer<typeof insertJobSchema>;
export type Job = typeof jobsTable.$inferSelect;